% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_png.R
\name{check_png}
\alias{check_png}
\title{Generate a test PNG}
\usage{
check_png(file_name)
}
\arguments{
\item{file_name}{file to write}
}
\value{
invisibly returns the plot
}
\description{
Generate a test PNG
}
