% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_png.R
\name{check_png_grid}
\alias{check_png_grid}
\title{Produce a plot using grid only}
\usage{
check_png_grid(file_name)
}
\arguments{
\item{file_name}{file to write}
}
\value{
invisibly returns the plot
}
\description{
Produce a plot using grid only
}
\examples{
check_png_grid("test.png")
}
